inerg-assessment-project/requirements.txt
black==24.10.0
blinker==1.9.0
click==8.1.7
et_xmlfile==2.0.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy-extensions==1.0.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
tomli==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
xlrd==2.0.1
